backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
